added functionality for user model

+++ controllers/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 //@desc Get all contacts
 //@route GET /api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Create contact
 //@route POST /api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req,res) => {
     console.log(req.body);
     const {name,email,phone} = req.body;
@@ -24,13 +24,14 @@ const createContact = asyncHandler(async (req,res) => {
         name,
         email,
         phone,
+        user_id: req.user.id
     });
     res.status(200).json(contact);
 });
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
 
@@ -38,12 +39,18 @@ const getContact = asyncHandler(async (req,res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User do not have permission to see other user contacts");
+    }
+
     res.status(200).json(contact);
 });
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
 
@@ -52,19 +59,29 @@ const updateContact = asyncHandler(async (req,res) => {
         throw new Error("Contact not found");
     }
 
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User do not have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(req.params.id, req.body, {new: true});
     res.status(200).json(updatedContact);
 });
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
 
     if(!contact){
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if(contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("User do not have permission to delete other user contacts");
     }
 
     await Contact.findByIdAndDelete(req.params.id);
